added lint and updated tags

+++ controllers/tagController.js
@@ -1,37 +1,5 @@
-const Tag = require("../model/tags");
-
-// const createtag = async (req, res) => {
-//   const { tag_name } = req.body;
-
-//   const tag = await Tag.create({
-//     tag_name: tag_name,
-//   });
-
-//   if (!tag) {
-//     return res.status(500).json({
-//       code: "Internal-Server-Error",
-//       error: "Something went wrong while creating tag",
-//     });
-//   }
-//   return res.status(200).json({
-//     message: "Tag created Successfully !!",
-//     Bank: savedBank,
-//   });
-// };
-
-// const getTag = async (req, res) => {
-//   try {
-//     const tags = await Tag.find({});
-//     if (!tags) {
-//       res.status(404).json({
-//         message: "no tag found",
-//       });
-//     }
-//     res.status(200).json({
-//       data: tags,
-//     });
-//   } catch (error) {}
-// };
+const Tag = require('../model/tags');
+const Statement = require('../model/statement');
 
 const createTag = async (req, res) => {
   try {
@@ -39,11 +7,10 @@ const createTag = async (req, res) => {
     const savedTag = await Tag.create({
       tag_name: tag_name,
     });
-    if (!savedTag)
-      return res.status(500).json({ message: "Internal server error" });
-    res.status(200).json({ message: "Tag created successfully", savedTag });
+    if (!savedTag) return res.status(500).json({ message: 'Something Went Wrong while Saving' });
+    res.status(200).json({ message: 'Tag created successfully', savedTag });
   } catch (err) {
-    console.error("Error creating tag:", err);
+    console.error('Internal Server error:', err);
     res.status(400).json({ success: false }, err.message);
   }
 };
@@ -51,44 +18,46 @@ const createTag = async (req, res) => {
 const getAllTags = async (req, res) => {
   try {
     const tags = await Tag.find();
-    if (!tags)
-      return res.status(500).json({ message: "Internal server error" });
-    res.status(200).json({ message: "tags data retrive successfully", tags });
+    if (!tags) return res.status(500).json({ message: 'No Data Found' });
+    res.status(200).json({ message: 'tags data retrive successfully', tags });
   } catch (err) {
-    console.error("Error creating tag:", err);
-    res.status(400).json({ success: false }, err.message);
+    console.error('Error creating tag:', err);
+    res.status(500).json({ success: false }, err.message);
   }
 };
 
 const updateTag = async (req, res) => {
   const id = req.params.id;
   const data = req.body;
-  // const { tag_name } = req.body;
-  // console.log(_id, tag_name);
   try {
+    if (!id) {
+      return res.status(404).json({ message: 'Tag ID is required' });
+    }
     const updatedTag = await Tag.findByIdAndUpdate(id, data, {
       new: true,
     });
-    if (!updatedTag)
-      return res.status(500).json({ message: "Internal server error" });
-    res.status(200).json({ message: "tags updated successfully", updatedTag });
+    if (!updatedTag) return res.status(500).json({ message: 'Something Went Wrong' });
+    res.status(200).json({ message: 'tags updated successfully', updatedTag });
   } catch (err) {
-    console.error("Error creating tag:", err);
-    res.status(400).json({ success: false }, err.message);
+    console.error('Error creating tag:', err);
+    res.status(500).json({ success: false }, err.message);
   }
 };
 
 const deleteTag = async (req, res) => {
-  const { tagId } = req.body;
+  const { tagId } = req.params.id;
   try {
+    if (!tagId) {
+      return res.status(404).json({ message: 'Tag ID is required' });
+    }
+    await Statement.updateMany({ tag_name: tagId }, { $set: { tag_name: null } });
     const deletedTag = await Tag.findByIdAndDelete(tagId);
-    console.log("Tag deleted:", deletedTag);
     if (!deletedTag)
-      return res.status(500).json({ message: "Internal server error" });
-    res.status(200).json({ message: "tags deleted successfully", deletedTag });
+      return res.status(500).json({ message: 'Tag deletion failed or tag not found' });
+    res.status(200).json({ message: 'tags deleted successfully', deletedTag });
   } catch (err) {
-    console.error("Error creating tag:", err);
-    res.status(400).json({ success: false }, err.message);
+    console.error('Error creating tag:', err);
+    res.status(500).json({ success: false }, err.message);
   }
 };
